adding Insert structure
A better way to handle insert or replacements. Each src token can have
an insert pointer which contains a vector of inserted tokens.

This allows to have a clean way to write the output without having to
renumber the main src vector and associated dbs. This is also moves us
closer to remove the test mode.