PCH checkpoint #6
Solved the inner namespace issue, for example JsonImpl which was being
left in the h file because it had no exported names.

The whole inner namespace block is detected and moved wholesale to the
cc file.

This gets us very close to done, we just need to deal with the top
namespace missing from the cc file.