DemandPagedMemory is now used in inflater.h
And therefore we come full circle to the adventure we started about 20
days ago.

The problem was that std::vector is not performant at all when it comes
to "deflate", because it has to come with lz77

for (int ix = 0; ix != count; ++ix)
output_.insert(output_.end(), output_[s++]);

that insertion causes lots of copies.

Along the way we have  had to invent
plx::Globals
plx::VEHHandler
plx::DemandPageMemory