fixed bezier's problem
d2d bezier has the p0 start point implicit and nanosvg no documentation
so I had made one geometry figure per bezier because I assumed that the
last point of the path segment was not the same as the first point of
the next path segment.
But closing the figure you need to decide if it is open or closed which
makes no sense since that is a property of the entire path, not of a
segment.

Assuming that all the segment in a path do not have gaps seems
reasonable and gives the correct result for the two tests svgs.