thight bounds on the svg geometry
now we request a surface just the size of the svg geometry so we need to
update the dc transform so that the origin is correct or else we end up
clipping the right and bottom of the svg.

We could require that the svg asset is set to be thight to the origin
but this code is not too bad.